Update to bsnes v032 release.
- Core: simplified CPU / SMP flag calculations
    - Added ALSA audio output driver to Linux port [Nach]
    - Improved font handling for Windows and Linux ports
    - Greatly cleaned up the user interface
    - Windows port now uses Unicode instead of ANSI
    - Added localization support
    - Config and locale files can now be placed inside bsnes executable directory for single-user mode, if desired
    - Fixed crashing bug with HQ2x on Linux/amd64 port [RedDwarf, Nach]
    - Hid "Power Cycle" option by default, as it is too similar to "Reset"
    - Slighty tweaked program icon [FitzRoy]
    - Minor code cleanups -- replaced union bitfields with templates, improved memory allocation, etc

+++ src/lib/libfilter/hq2x.cpp
@@ -71,15 +71,15 @@ void HQ2xFilter::render(
   uint32_t *out0 = output;
   uint32_t *out1 = output + outpitch;
 
-  #define W1 input[-1 - pitch]
-  #define W2 input[ 0 - pitch]
-  #define W3 input[+1 - pitch]
+  #define W1 input[-1 - (int)pitch]
+  #define W2 input[ 0 - (int)pitch]
+  #define W3 input[+1 - (int)pitch]
   #define W4 input[-1]
   #define W5 input[ 0]
   #define W6 input[+1]
-  #define W7 input[-1 + pitch]
-  #define W8 input[ 0 + pitch]
-  #define W9 input[+1 + pitch]
+  #define W7 input[-1 + (int)pitch]
+  #define W8 input[ 0 + (int)pitch]
+  #define W9 input[+1 + (int)pitch]
 
   input += pitch;
   memset(out0, 0, 2048); out0 += outpitch << 1;
